// src/renderer/lib/components/Component/Component.d.ts
/**
 * Base component
 */
export default class Component {
  /**
   *
   * @param {*} parent - CHERRIES Component - if undefined - get app form
   */
  constructor(parent: any);
  uid: string;
  baseClass: string;
  designClass: string;
  cont: any;
  set parent(arg: any);
  get parent(): any;
  changeOrig: boolean;
  onEvents: string[];
  eventListeners: any[];
  _data: {};
  prevAnchor: string;
  set top(arg: any);
  get top(): any;
  set left(arg: any);
  get left(): any;
  set width(arg: any);
  get width(): any;
  set height(arg: any);
  get height(): any;
  boxSizing: string;
  type: any;
  kind: string;
  addResizeListener(): void;
  _armStyles(): void;
  stylesUnitConvert: string[];
  styles: string[];
  properties: string[];
  _redefineStyles({ styles, targetElement, doDelete, unitConvert }: { styles: any; targetElement: any; doDelete: any; unitConvert: any }): void;
  _redefineProperties(properties: any): void;
  set design(arg: any);
  get design(): any;
  _design: any;
  set designSwitch(arg: any);
  get designSwitch(): any;
  destroy(): void;
  destroyChildren(el: any): void;
  muteChildren(op: any): void;
  set onMouseWheel(arg: any);
  set onMouseDown(arg: any);
  set onMouseMove(arg: any);
  set onMouseLeave(arg: any);
  set onMouseEnter(arg: any);
  set onMouseUp(arg: any);
  set onDblClick(arg: any);
  set onClick(arg: any);
  set onDragStart(arg: any);
  set onDragEnter(arg: any);
  set onDragOver(arg: any);
  set onDrop(arg: any);
  set onTouchStart(arg: any);
  set onTouchEnd(arg: any);
  set onTouchMove(arg: any);
  set onTouchCancel(arg: any);
  setAttribute(a: any, v: any): void;
  getAttribute(a: any): any;
  removeAttribute(a: any): any;
  getKeys(e: any): {
    code: any;
    key: any;
    keyCode: any;
    repeat: any;
    type: any;
    altKey: any;
    ctrlKey: any;
    shiftKey: any;
    metaKey: any;
  };
  getKeysAndMouse(e: any): {
    event: any;
    target: any;
    state: any;
    rect: any;
    contRect: any;
    uid: any;
    type: any;
    name: any;
    pos: {
      x: number;
      y: number;
    };
    button: any;
    altKey: any;
    ctrlKey: any;
    shiftKey: any;
    metaKey: any;
  };
  set name(arg: any);
  get name(): any;
  set userClass(arg: any);
  get userClass(): any;
  set enabled(arg: any);
  get enabled(): any;
  set draggable(arg: any);
  get draggable(): any;
  storeStyle(el: any, style: any): void;
  storedStyles: any[];
  restoreStyle(el: any, style: any): void;
  setStyle(el: any, style: any, v: any): void;
  set text(arg: any);
  initialText: any;
  set visible(arg: any);
  get visible(): any;
  display: string;
  _setListener(type: any, func: any): void;
  _clearListener(node: any, type: any): void;
  set ignoreEvents(arg: any);
  get ignoreEvents(): any;
  show(): void;
  hide(): void;
  _getParentInstance(parent: any): any;
  _parent: any;
  origTop: any;
  set bottom(arg: number);
  get bottom(): number;
  origBottom: number;
  origLeft: any;
  set right(arg: number);
  get right(): number;
  origRight: number;
  set rect(arg: any);
  get rect(): any;
  _rect: any;
  set align(arg: any);
  get align(): any;
  position: string;
  set anchor(arg: any);
  get anchor(): any;
  _resetPosition(): void;
  anchorUpdate(): void;
  _onResizeCB(): void;
  center(): void;
  centerH(): void;
  centerV(): void;
  set devhint(arg: any);
  get devhint(): any;
  _devhint: any;
  set hint(arg: any);
  get hint(): any;
  set flash(arg: any);
  get flash(): any;
  _tempColor: any;
  _tempBackgroundColor: any;
  _counter: number;
  color: any;
  backgroundColor: any;
  flashInt: NodeJS.Timer;
  set autotranslate(arg: void);
  get autotranslate(): void;
  refresh(): void;
  getPos(e: any): {
    x: number;
    y: number;
  };
  remove(): void;
  set state(arg: any);
  get state(): any;
  getState(e: any): any;
}